src/e/edit/Diffable.java: make it possible to explicitly declare (and later query) the FileType for a Diffable.
src/e/edit/CompareSelectionAndClipboardAction.java: 
src/e/edit/ETextWindow.java: supply the appropriate FileType when creating Diffables.

src/e/edit/SimplePatchDialog.java: and use the FileType to configure the PTextArea that's showing the patch. This ensures we get the appropriate spelling exceptions. We also get language-specific coloring; we didn't reliably get it previously because we were trying to guess the file type from the possibly-temporary filenames, which don't have a useful extension, and from the patch content, which isn't generally recognizable as the source language.