www/downloads.html: My existing Terminator and Evergreen worked fine after upgrading to Mac OS 10.6.  A downloaded Terminator also worked, modulo the allegedly damaged image file.
lib/build/universal.make: Building on 10.6 required two changes, one to work around what looks like a minor regression in the PowerPC support and the other to work around the default compiler being changed to a less prehistoric version of gcc.  Teh interwebs lead me to think that 4.0.1 was the gcc version on previous Mac OSs, so I don't *think* these changes will break Ed's nightly build on 10.whatever.  Unexpectedly, then, it looks like we can, if we choose, support 10.4, 10.5 and 10.6, building on any of them.