src/terminator/model/TextLine.java: A minimal fix for the NPE in getStyledTextSegments discussed in http://groups.google.com/group/terminator-users/browse_thread/thread/e2c4b2a69eb6d467#.
We weren't initializing the styles from offset + count to styles.length.  Examination of the history shows that this method was originally only used when the overwrite was at the end of the existing line (making the uninitialized region empty).

This problem is reproducible, in a new window, by displaying some text, moving left, then overwriting some, but not all, of it:

echo $'hello\E[4D\E[35mwor'

That isn't reproducible in a window which has undergone any Style changes, even if you prefix the command with "reset; ".  This will prevent anyone with eg a colored prompt from reproducing the problem.

The reason for this is that, despite now carefully calling equals() instead of == on Style, we're using the default definition for equals(), which, as I understand it, basically is ==.  This means that we can never return to a Style that equals the default style.  This in turn means that all of the code in this class that attempts to optimize the case of default-styled lines ceases to have any effect once the current Style is changed, which will surely happen in many, many terminal windows.  We're littering memory with equivalent Styles that our code doesn't realize are equivalent.

Defining Style.equals() would at least reactivate the optimizations here.  Elliott instead suggests doing something to avoid holding on to multiple equivalent Style instances in the first place in:

http://groups.google.com/group/terminator-users/browse_thread/thread/cda99774ddc19b07?pli=1

How would we do that?  The first method that occurs to me is to cache the Style instances that we've used in a container, then have makeStyle look them up.  HashSet springs to mind as a container.  That would, I think, require that we implement equals and hashCode.  So perhaps I should check-in my equals, though I don't have a hashCode and I'm wondering why we don't have more definitions of equals and whether there's some trick for defining tuple equality.

But, before that, Elliott had was some belt-and-braces ideas that I should consider here.