src/terminator/model/Style.java: Define the equals method that we now diligently call, such that our careful code in TextLine to optimize for the default style is activated when we return to the default style by means other than the Reset menu item.
A working equals made it easier to reproduce the null pointer exception in TextLine that was fixed last week.  Even once we stop hanging on to equal Style instances, we'll need this to to determine that a requested instance is already in the weak hash set.  At the point where we implement hashing, we'll want to revisit the deliberately lame hash function here.  A toString was useful to me in discovering this omission...

src/terminator/model/StyledText.java: ... and the getDescription that I cribbed it from isn't called anywhere.  Anyone wanting it in future can now call getStyle().toString() (or do something more complete).

Despite leaving the email list with the impression that I've been running with this all week, I've actually just knocked it up, with one eye on Elliott's summary of Bloch's guidelines.  So it's not only My First Equals but not tested with much more than the compiler.

I can't help thinking that we might be better off splitting Style into StyleChange and CompleteStyle.  The former would get the Boolean fields and wouldn't require equals.  The latter would have boolean fields.  It would still need equals but at least the implementation wouldn't be quite so cumbersome.  You'd apply a StyleChange to a CompleteStyle to get another CompleteStyle.