src/terminator/view/highlight/FindHighlighter.java: There's only one call to new Style outside Style, so it's to take one step towards having Style manage its own instances...
src/terminator/model/Style.java: ... at the zero expense of moving the autoboxing (if that's the right word for the conversion from "boolean isBold" to "Boolean isBold") from the Style constructor to the makeStyle class method.  I was hoping that this would let me make the Style constructor private, as Elliott had desired in the discussions about the 256 color patch.  The compiler then pointed out two places where we "extends Style" and override its methods for getting eg the colors.  Well, that presumably makes a partial nonsense of my previous check-in.  Given that we want dynamically varying Styles, do we not want Style to become an interface?  I'm imagining a FixedStyle which implements that interface in terms of two Color fields and three booleans.  I settled for "protected" in the meantime, partly to hint to any similarly naive future maintainer that there are subclasses.