lib/build/dist.make: The parts of universal.make that know about the directory structure on jessies.org and particularities of the uploaded filenames.
lib/build/universal.make: This is then delegated to via a variable, DIST_MAKEFILE, which can be set by the calling code in order to provide a different implementation of "make native-dist", for uploading elsewhere than jessies.org.  While DIST_MAKEFILE is required to be a makefile fragment, its implementation need not be more than a couple of lines, to delegate to a script, passing on the values of any variables that the script might find useful.