lib/build/universal.make: Cope with find-jdk-root.rb returning the location of an Apple-style JRE, which includes all the JDK stuff, just not with the standard layout.  Doing this arguably improves the code here, especially by encouraging us to defer the test for jni_md.h until we can implement it more elegantly.
lib/build/package-for-distribution.rb: Only use #{find_jdk_root()}/bin/jar where jar isn't on the path, which will only be Windows, thus avoiding it on Mac, where bin/ no longer exists...

bin/find-jdk-root.rb: ... after abandoning the special-casing here which caused trouble in the previous couple of revisions.  I'm not really convinced that this is a good idea, but I'd made the universal.make change before I realized that the problem could more easily be fixed in the way I committed just now.  Even if we were to decide to revert the change to this file, most of the universal.make change seemed worth keeping.