src/terminator/TerminalLogWriter.java: Elliott right suggests leaving a code-comment to explain why canWrite failed us.
(1614/3227/1.6.0_20-b02/Windows XP 5.1 Cygwin 1.7.5/x86 x1)
(1615/3227/1.6.0_20-b02/Windows 7 6.1 Cygwin 	1.7.5/x86 x4)

I was trying to connote "I don't care" by passing a null suffix, having seen the documentation seemingly encouraging that, but I did think that it's then not self-evident what that argument means and it clearly set one reader wondering whether there was a lower limit on the size of the suffix.  I can't imagine it failing with an empty string but ".tmp" makes it clearer what the author thought the parameter was for.