native/all/pty/PtyGenerator.h: Sean McNeill reports:
If I use new shell or new shell tab, I get an error:

java.io.IOException: Could not start process "/bin/bash".
    at terminator.terminal.PtyProcess.<init>(PtyProcess.java:129)

I know that forwarding the fork failure in the manner we were doing is traditional, but it's denying us the errno that might go a little way towards explaining Sean's issue.

src/terminator/terminal/PtyProcess.java: We'll no longer assign -1 to pid, so we'd either have to initialize it to that or lose the check here that will no longer be reached.  We'll lose the name of the executable but, for a fork() failure, the executable hasn't come into play.  Failure to find/exec the executable is still reported cleanly in the new tab/window.