bin/invoke-java.rb: Try to diagnose Jairo Jacome's issue with cygpath failing, which doesn't occur when run from "a Cygwin window", thus rending the stderr output pointless.  I suspect this is a Cygwin rebase()ing issue.
Switch from the Java/C++ "throw" to the "raise" that we use elsewhere in our Ruby.  I don't really remember the reason and I'll doubtless forget it again, if anyone bothers to remind me, but I've learned that "raise good, throw bad" in other projects.  "throw" was doing the job here, so this is a needless risk.