lib/build/per-directory.make: Delegate the linker flags so they can be conditional on the compiler version...
lib/build/universal.make: ... so we can use the appropriate flags for the newer mingw packaged for Cygwin.

I haven't switched to this new compiler and release as, while the resulting binaries successfully started Terminator on Cygwin-1.7.9 from rxvt, they did not do so from the desktop shortcut.  Today's DLL snapshot, of several further months of development, including ChangeLog entries relating to DLL startup, suggested that the failure in future will be more nicely presented but also more consistently presented: it happened even from rxvt.  Today's upload was built with the old compiler but the 1.7.9 release.  It also doesn't work (even on 1.7.9 - the xpg_strerror_r thing stops it on 1.7.7).  I've tried but failed to compile the cygwin DLL myself.  I had to install g++-4 to compile something locale-related but that went fine until some "mkexport" stage, when objcopy and perl seemed to get stuck in a loop for hours.  I wanted to be able to compile myself so I could find where exactly the change in behavior that broke us went in.