src/terminator/Terminator.java: Fix a probably long-standing problem where we don't quit on Windows once the last frame has gone. A day of experimentation, some of which I recorded at https://groups.google.com/d/msg/terminator-users/8UKYHEohUzo/TygEwsE1DA0J, led me to conclude that Terminator was doing something out of "global constructors" that was too early for the AWT infrastructure, seemingly not on Linux but on Windows.  This caused the creation of a second EventQueue thread, which I don't see after this change.  One EventQueue thread disappeared more or less with the final frame, but the other remained.
Defer the construction of the singleton Terminator object from "global constructors" to, I think and hope, the EDT, via main()'s EventQueue.invokeLater().  I've switched to the orthography used for eg Evergreen and e.edit.Advisor.  For the order of "synchronized" and "static", I surveyed the JDK, having got no clear answer from the jessies code.

While we might not be doing AWT stuff directly in this class, the dependencies ripple out.  We use GuiUtilities and that starts EventDispatchThreadHangMonitor and that seems pretty AWT-related to me.  Perhaps the crux of the problem was elsewhere: I stopped when I found the first solution, as it seemed in keeping with Evergreen.