lib/build/googlecode_upload.py: Goodbye googlecode download uploader, we hardly knew ye.
lib/build/drive.rb: Hello Google Drive API example code, lightly hacked to upload a binary into a shared subdirectory of my Google Drive, if it doesn't already exist there and assuming that you've signed up for the Google API Console and created a Project and a "Client ID for native application" and done "Download JSON" to client_secrets.json in this directory, then, the first time you run it, which has to be interactively, agree to the access, which I must have granted to your Google account, which I can't until you tell me which account I should grant it to.  There's a load of debugging output in here and it may not work at all, as is the way of the first checked-in revision.

lib/build/dist.make: Remove the support for the old googlecode.com, as that seems unlikely to be revivified, in favor of the new.  I couldn't work out a download URL from the source filename, so the checking whether we've already uploaded needed delegation to the API-using script.