lib/build/drive.rb: The API depends on "Launchy", which doesn't work for URLs at least on Cygwin64, and doesn't look trivially remedied, at least not for URLs that are long and/or require escaping.  This can be trivially demonstrated with:
ruby -we 'require "launchy"; Launchy.open("http://www.google.com/")'

Which opens a cmd window containing:

ERROR: No application found to handle '/b'
Try `launchy --help' for more information.

C:\cygwin64\lib\ruby\gems\1.9.1\gems\launchy-2.4.2\lib\launchy\applications>

Unless you apply:

--- browser.rb.orig     2014-03-31 15:50:43.347614500 -0700
+++ browser.rb.martind  2014-03-31 15:51:47.305148500 -0700
@@ -17,7 +17,7 @@
     end

     def cygwin_app_list
-      [ 'cmd /C start "launchy" /b' ]
+      [ 'cmd /C start /b' ]
     end

     # hardcode this to open?

That's not enough to sort out a Drive authentication URL, which fails with "The system cannot accept the START command parameter".  By enabling debug, we get to see the URL, which the user can copy-and-paste and away we go.  It's a one-time thing, so not anything like as huge a deal as it was to work out what was wrong.