native/all/update-login-record/updateLoginRecord.cpp: The code for nativeUpdateLoginRecord from Terminator.  This was working fine as-was on Cygwin, where utmpx doesn't require root privilege but pulling it out into its own translation unit that doesn't depend on JNI...
native/Headers/updateLoginRecord.h: ... needn't stop us calling it via a C++ interface from the JNI.  I don't propose to do that.  Instead...

native/all/update-login-record/update-login-record.cpp: ... here's a command-line interface to the same...

native/Headers/parseInteger.h: ... for which I want an atoi-alike that just gives me a bool to say if it worked and, as far as the interface is concerned, works with any size integer I give it, though I haven't bothered to make the implementation do the right thing with negative arguments, as I don't need them for this.

lib/build/per-directory.make: To give us something extra over calling the code directly from JNI, we need the executable to be setuid root.  This will require password-free sudo to build.  I'm not sure what will happen without that.  The hack for determining which executables should be setuid root won't survive an encounter with a second example but that might well have been the case with any of the alternatives I considered and this one was the smallest and most self-contained.

native/all/update-login-record: I don't understand Subversion.