[Bug] D111633: Java error hung my Terminator windows when copying contents of a Terminator window to a gedit window
src/e/util/LazyStringSelection.java: http://docs.oracle.com/javase/7/docs/api/java/awt/datatransfer/Transferable.html says that getTransferData is expected to throw IOException "if the data is no longer available in the requested flavor".  Comments in this file already discuss the case where the selection is very large.  If we run out of memory constructing the text, then we were ending up throwing OutOfMemoryError.  I suspect that doesn't get caught by some of the toolkit code that perhaps ought to catch it because we end up not processing any GUI events, with only one interesting thread:

   "AWT-XAWT" daemon prio=10 tid=0x00007f6064252800 nid=0x605a waiting on condition [0x00007f604903e000]
      java.lang.Thread.State: TIMED_WAITING (parking)
               at sun.misc.Unsafe.park(Native Method)
               - parking to wait for  <0x00000007002b3aa8> (a java.util.concurrent.locks.AbstractQueuedSynchronizer$ConditionObject)
               at java.util.concurrent.locks.LockSupport.parkNanos(LockSupport.java:226)
               at java.util.concurrent.locks.AbstractQueuedSynchronizer$ConditionObject.await(AbstractQueuedSynchronizer.java:2176)
               at sun.awt.SunToolkit.awtLockWait(SunToolkit.java:265)
               at sun.awt.X11.XlibWrapper.XNextSecondaryLoopEvent(Native Method)
               at sun.awt.X11.XToolkit.run(XToolkit.java:530)
               at sun.awt.X11.XToolkitThreadBlockedHandler.enter(XToolkitThreadBlockedHandler.java:49)
               at sun.awt.datatransfer.DataTransferer.convertData(DataTransferer.java:2351)
               at sun.awt.X11.XSelection.convertAndStore(XSelection.java:488)
               at sun.awt.X11.XSelection.handleSelectionRequest(XSelection.java:565)
               at sun.awt.X11.XSelection.access$500(XSelection.java:45)
               at sun.awt.X11.XSelection$SelectionEventHandler.dispatchEvent(XSelection.java:747)
               at sun.awt.X11.XToolkit.dispatchEvent(XToolkit.java:489)
               at sun.awt.X11.XToolkit.run(XToolkit.java:584)
               at sun.awt.X11.XToolkit.run(XToolkit.java:505)
               at java.lang.Thread.run(Thread.java:745)

Rethrowing that with an IOException wrapper causes the middle-button paste into another application to do nothing but stops Terminator from locking up.

src/e/debug/EventDispatchThreadHangMonitor.java: We'd have solved this yesterday if my grep for OutOfMemory in the terminator log had found a hit.  Sadly, that exception was only being reported on the console of the process that invoked terminator.  That parent was a gnome-panel icon so, if the error was recorded anywhere, it was somewhere I didn't think to look.  If we catch the error here, then we can log it.  While we're not allowed, by the signature, which I dare not change, as it's an @Override, to rethrow Throwable here, we can rethrow Error.  Rethrowing it causes it to be reported on the invoker's console too but I'm scared by the thought of claiming to handle everything.