[Bug] D118239: Java NullPointerException thrown when attempting "revisiontool leak.rb"
salma-hayek/src/e/util/Log.java: Defer the addition of the Java and OS banner...

salma-hayek/src/e/gui/AboutBox.java: ... and the version information until we get a real warning, in the hope that...

evergreen/src/e/edit/ShellCommand.java: ... we might never get such a warning, particularly for tool invocations rather than heavyweight applications, and so better adhere to the Unix bushido of silent success.  If we don't see any output, then we don't need the clutter of an empty errors window.  It's somewhat dislocating to get the errors window popping up later, particularly when the trigger is something that we don't care much about.  Then again, that's an incentive to fix bogus warnings like some of the instances of:

java.lang.IllegalArgumentException: fromRevision must be newer than toRevision or this code needs fixing