terminator/src/terminator/Terminator.java: Any syntax error has already been reported...
salma-hayek/bin/invoke-java.rb: ... and behaving as if we have a bug or Java has crashed just obscures that, making eg Yogesh think that there was some secondary failure.

This is distressing duplication of a magic number, but we don't seem to System.exit() for anything other than 1 or EXIT_FAILURE anywhere else.

We already have a code path for exiting with clean failure for non-GUI programs, but there's no obvious back-channel to say that we didn't get to the GUI part yet.  If we didn't get that far because Java crashed, as it is wont to do on Cygwin, then we'd want the full monty report.

I was tempted to just return success from Terminator.java, on the grounds that, if TerminatorServer is running, the same syntax error from terminator -e causes the terminator script to exit with success.  I concluded that that's another infelicity.