salma-hayek/lib/build/universal.make: Pass -m32 to the C/C++ compiler and linker if we're asked to build for TARGET_ARCHITECTURE=i386 when target-os.rb says our native architecture is amd64.
salma-hayek/lib/build/swiftboat.sh: That way, I abandon another dependency on my Squeeze chroot, with its annoyingly noisy alien (per the previous check-in) and its ancient default Ruby (per the check-in before).  This is untested and I've tried to be a bit fly with my bash, to get closer to saying one thing per line, but I wanted to check it in together with the somewhat tested universal.make change because this change is the reason for that one.