[RFE] D129126: DRR tests which reboot the system should use bounceDevHnas
evergreen/src/e/edit/ExternalTools.java: After some puzzlement, I don't think FileIgnorer is following symlinks at all.  I think the accessor is just controlling whether symbolic links are returned at all, so rename it...

evergreen/src/e/edit/FileIgnorer.java: ... while refining the default behavior, that used by Evergreen's WorkspaceFileList, to satisfy my goal today:

We have some Ruby code which is sometimes used as a library and sometimes as an executable.  When used as a library, the script is often most naturally named for a function, so in camelCase.rb.  When used as an executable, the script is often most naturally named in kebab-case.  Sometimes the executable uses the library, but sometimes it just is the library, via a relative symlink.  The absence of the kebab-case name in Evergreen's file list caused one of our users to think that a usage error message that used that case was wrong.

With this change, we get to see the kebab-case file name and the camelCase.rb one.  But what about the worry that we'll have two unsynchronized views on the same file?  That's taken care of, by canonicalization, in the opening code already, where we might get a symbolic link that's not from the workspace file list.

Lucky I tested it, though, or I'd have been bitten by Java's string equality gotcha and added thousands of uninteresting header file symlinks to all of my BlueArgh "fish" workspaces.

IOException is as annoying as ever.  Do I ignore it here, which I'd kick myself for if I then spent ages debugging, or do I potentially spam a log with thousands of warnings?  I want to leave any such exceptional case to the caller, like, you know, it's an exception!  I found somewhere else where we turn the IOException into a RuntimeException.  That'll do me.