salma-hayek/lib/build/dist.make: salma-hayek/lib/build/drive.rb: I don't especially want to waste bandwidth, time and storage by uploading the installers twice... but I don't currently see how I'll be able to make stable links that work by name rather than id, compelling me to use drive.files.update and send the content a second time.  I could forget about keeping the old versions, relying on Drive's file history, but that doesn't give me an obvious way to record the version numbers and ties us more closely to the Drive way of doing things.
I could manually copy the most recent installer of each type to the hopefully stable name for the latest version, thus giving me a way to get the ids, for linking to from the github wiki, but I plan to let the code programmatically create them instead, not least so that I have a record of which platforms it's worked on and so it doesn't give the appearance that eg a Solaris installer was built in 2017.  In case I can't get a stable link to the latest version of each installer and in the meantime, I've told Drive to put the stably named copy of the latest versions in a folder that isn't cluttered by all the old versions.  I'll surely be able to link to that directory.

I've augmented the commentary on how I set this up.