salma-hayek/lib/build/package-for-distribution.rb: I presume I've been getting this every day for years without noticing:
/home/martind/jessies/nightlies/salma-hayek/lib/build/package-for-distribution.rb:207: warning: File.exists? is a deprecated name, use File.exist? instead
/home/martind/jessies/nightlies/salma-hayek/lib/build/package-for-distribution.rb:383: warning: File.exists? is a deprecated name, use File.exist? instead
/home/martind/jessies/nightlies/salma-hayek/lib/build/package-for-distribution.rb:414: warning: File.exists? is a deprecated name, use File.exist? instead

Those were the only such matches in the code base.  There are other calls to File.exist? even in the same source file.  At least consistency is an ethos, even if the grammar seems to be... unusual.  I traced the reasoning back to Matz and 2001 with http://blade.nagaokaut.ac.jp/cgi-bin/scat.rb/ruby/ruby-talk/18951.