terminator/src/terminator/model/TerminalModel.java: Don't add characters to the TerminalModel if there are so many of them that the display size would be more than 32768 pixels wide (which java can't cope with).
terminator/src/terminator/view/TerminalView.java: If we're trying to display a bit of a terminal close enough to the right-hand side of a 32768 pixel wide thingy, then show a warning that the text has been truncated.

In combination, this should sort out Martin's problems of Terminator locking up when he accidentally cats a massive file with no linefeeds. Indeed, I've experienced this too, and it's not fun. Given that Java can't show anything over 32768 pixels horizontally anyway (at least, not with any reliability), and given that if your display is this big, losing a few bytes is the least of your worries, this should solve the expensive horizontal lines problem in a nice, usefully simple way.