terminator/lib/terminfo/terminator.tic: A few tweaks derived from the Debian version of the terminator terminfo file. * Remove km (we don't do meta key -> 8th bit char molestation) * Remove the duplicate smacs * Add sgr and augment sgr0
salma-hayek/lib/build/universal.make: Now that the above fixes are in place in our terminfo file, it parses correctly on FreeBSD, so there's no need to disable it. Leave a comment behind though, so in case someone needs to disable it on some platform in the future, they don't have to expend as much time as I did.