terminator/src/terminator/TerminatorFrame.java: Only calling setWindowOpacity when the alpha value changes wasn't enough to keep transparency working for me on Java 8 on Linux.  I didn't look into that further, instead choosing to investigate Elliott's idea from https://github.com/software-jessies-org/jessies/issues/25#issuecomment-427723198 of overriding isUndecorated.  That didn't work in the obvious implementation, for a reason I'll leave in a Log.warn in GuiUtilities.setFrameAlpha.  As a Swing app, we're single-threaded in our accesses to the UI, right?  For the price of a little bit of sick in the mouth, we can override isUndecorated temporarily, which did work for me.
salma-hayek/src/e/util/GuiUtilities.java: Do I duplicate the exception handling for the new call to setOpacity or do I abstract it out?  That all seems like a lot of effort to go to for transparency support.  I was using it on the first generation Intel MacBook that is the reason we're still only relying on Java 6, but I haven't used that machine in months, partly because it's dying.  Nowhere else would I use anything less than Java 8.  Perhaps, then, we don't need to keep all the clever code from http://elliotth.blogspot.com/2007/08/transparent-java-windows-on-x11.html, which I could only re-read yesterday via the Wayback Machine.  I haven't tried this on Windows but the old code never worked for me there either, since my last physical machine with Windows shuffled off this mortal coil with the end of support for XP.