Clarify the rt.jar warnings.
I'd already accidentally checked in a change to use $(warning) rather
than $(error), and it looks like that's going to increasingly be the default.
I couldn't work out how to get Java 8 on macOS 11, for example. (Even more
worryingly, Java 11 doesn't even have an rt.jar, so it's likely that when we
bump the "lowest supported LTS" up from 8, whatever we bump it to won't have
an rt.jar, and not all of our platforms will have that LTS release.)

But we can at least clarify why we care about rt.jar, and what you
should think about if you're seeing the rt.jar warning and are not
in a position to just install an old JDK.

For Debian, we might want to give the copy & paste `sudo apt` command
you need to install Java 8? But this seems like a step forward
anyway.