cope with update-login-record being missing (#46)
terminator/src/terminator/terminal/PtyProcess.java: If the user doesn't grant us sudo rights when we're building update-login-record - and why should they type their password into our code? - then I can see Make getting stroppy and removing the binary, especially as we add the chmod command to LDFLAGS(!):

https://github.com/software-jessies-org/jessies/blob/f0ffa063b1cc255cfd2f984007e9e1dfec9bc7dd/salma-hayek/lib/build/per-directory.make#L125

If there's no executable, then findSupportBinary is going to return null.  We surely don't intend to stop Terminator working when we've failed to update this nice-to-have login record business.  I'm not sure it's even worth logging a warning but let's not silently hide errors without an explicit argument that it's worth it.