salma-hayek/lib/build/drive.rb: Abandon compatibility with google-api-client 0.8.7, the last version that worked with Ruby 1.9.3.  Cygwin, which I don't use any more, has Ruby 2.6 and Debian Stretch, on my current build server, has Ruby 2.3.3, letting me use google-api-client 0.32.1.  This lets me toss out some obsolete code and gets an interface that fits better with Ruby.  I took a minimal approach - changing just what was necessary, having failed to find an example of similar code or, indeed, any particularly good documentation.  I know I scoffed before at the idea of needing ri to read documentation on a Google product but it does seem the best choice here.  This v2 Drive API is long since deprecated.
I don't have a strong reason for needing this.  I did just commit a change to work's install-gems script to enforce google-api-client 0.8.7's dependency on activesupport < 5.  I'd like to get rid of that again.  I realized, though, while writing the check in comment, that I only needed it because I'd uninstalled activesupport without uninstalling google-api-client, so I didn't need it anyway.  Still, the code-comment about a chroot is misleading, as I haven't used that since 2017.  Hopefully upgrading from a version from 2016 to one from 2019 will stave off having to revisit this for longer.  Indeed, I see that, on Debian Buster, which I could use without formality, google-api-client 0.53.0 from 2021-01-21 works fine with this code.  That claims to be deprecated in favor in google-apis-drive_v2, which depends on google-apis-core, whose oldest version requires Ruby >= 2.4, which won't fly on Stretch.  Looks like that's a drop-in replacement, though, on Buster, so we're up to the minute, at least with v2, not v3, of "the simple REST clients for Google APIs", which aren't apparently the preferred interface.  Whatever.