terminator/src/terminator/terminal/TerminalControl.java: Expose CHARSET_NAME because, if you wanted to customize this for your Psion (wow), you presumably wanted everything stuffed into and out of that pipe treated in the same character set, rather than in the default one, surely UTF-8, where you're running Terminator.
terminator/src/terminator/terminal/escape/XTermEscapeAction.java: Fix the whitespace departures from local convention, far as I was able to make that out.  Log something when the reason it doesn't work is because you (ahem) forgot to set the preference.  Use break when we haven't understood but return when we have, for the benefit of seeing or not seeing the warning that follows the switch.

I found somewhere else in Terminator that was setting the selection and it was passing the StringSelection as the clipboard owner.  I don't know that whoever did that understood better than Simon's null but my "this" wasn't a ClipboardOwner, which suggests they did.

The test case was advertised with hello but hell didn't work:

printf "\033]52;p;$(printf "%s" "hell" | base64)\a"

... upchucking from:

java.lang.IllegalArgumentException: Input byte array has wrong 4-byte ending unit
	at java.base/java.util.Base64$Decoder.decode0(Base64.java:736)
	at java.base/java.util.Base64$Decoder.decode(Base64.java:538)
	at terminator.terminal.escape.XTermEscapeAction.perform(XTermEscapeAction.java:79)

... because the last character was being dropped from the sequence.  I guess Simon wasn't expecting the string terminator to have been dropped by the constructor, but, despite the FIXME about that, I'm glad that it has been, because, as can be seen when searching https://invisible-island.net/xterm/ctlseqs/ctlseqs.html for ESC \, there are perhaps two alternative string terminators to BEL, though we don't seem to implement that.  Better, I reckon, to stick to xterm's pt, pc, pd names, then it's clear what our intention is.

I implemented a little more of the support for X11's bewildering variety of selections.  I didn't do this because I want it but more to indicate that we're not a full implementation of the spec.  I'd forgotten what our stance on these selections was but it seems that we do offer some support.  I don't know whether we'll ever see p on its own or whether that's really the middle button selection but that's what I bound it to.

I implemented the symmetric amount of the support for sending the local clipboard or selection to the remote party.  My guess as to what the intended behavior was supposed to be from the above documentation was contradicted by trying it with:

xterm -xrm "xterm*allowWindowOps: true"

... which festooned its entry in the remote input buffer with the same sort of escape sequence as was parsed.  That's consistent with the documentation and I had been wondering how the remote end was otherwise supposed to know where the base64 string ended.  xterm defaults, as it says, to s0 but my understanding of that is hazy, so I default to c, because that I think I do understand, the clipboard is what Simon used and what's available on other platforms.

Perhaps it would have been more conventional to give String::split a -1 limit but I thought I'd save it the extra work once it had found a parse error.  Ugh, checked exceptions.  I cribbed the handling code from a similar situation elsewhere in the Jessies code.  I failed the Java interview question about == so might have done something similarly crass elsewhere, though at least I caught that with testing.