Fix the long-standing bug where sometimes the undo status of a file would erroneously suggest the file is dirty. This was triggered any time someone used undo, and then made new changes which caused the undo buffer to be truncated. This was an attempt to handle the case where someone writes stuff, saves the file, then undoes one or more operations before making new changes (thus truncating the undo buffer and making it impossible to return to the saved state). We now check whether the 'clean' state is after the current position in the undo buffer, rather than just checking if the undo buffer is being truncated.
It is an open question as to whether we should instead go the emacs route, and rather than truncate the undo buffer, invert the undone edits and append them. Both options are valid, but I'm just going with the simple fix for now as the current behaviour has never bothered me _too_ much.

This change also adds a new keyboard action to be able to dump the undo buffer status to the log when hitting alt-z. Not usually useful, but very much so when trying to debug the undo logic.